[UPD] - Update function parameters and variable types in SearchViewModel and SearchScreen to enhance code clarity and ensure compliance with Explicit API mode.

+++ android/app/src/main/kotlin/com/jabook/app/jabook/compose/feature/player/controller/AudioPlayerController.kt
@@ -262,7 +262,7 @@ public class AudioPlayerController
                             // Wait for controller with timeout
                             val controller =
                                 mediaControllerFuture?.get(
-                                    MediaControllerConstants.DEFAULT_TIMEOUT_SECONDS,
+                                    MediaControllerConstants.DEFAULT_TIMEOUT_SECONDS.toLong(),
                                     TimeUnit.SECONDS,
                                 )
                             mediaController = controller
